- update: using QR to add Gateway clients and gateway servers - fixed issues with Gateway client Ids

+++ app/src/main/java/com/example/swob_deku/Models/GatewayClients/GatewayClientDAO.java
@@ -23,7 +23,7 @@ public interface GatewayClientDAO {
     int delete(GatewayClient gatewayClient);
 
     @Query("SELECT * FROM GatewayClient WHERE id=:id")
-    GatewayClient fetch(int id);
+    GatewayClient fetch(long id);
 
 //    @Query("UPDATE GatewayClient SET projectName=:projectName, projectBinding=:projectBinding WHERE id=:id")
 //    void updateProjectNameAndProjectBinding(String projectName, String projectBinding, int id);
